enable Fn::Transform to use CFn macros
